refactor: rename logo file
