+++ frontend/src/components/app/Footer.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import bmjvLogo from "~/assets/img/BMJV_de_v1__Web_farbig.svg";
+import bmjvLogo from "~/assets/img/bmjv-de-v1-web-farbig.svg";
 
 const route = useRoute();
 const hideFeedbackForm = computed(() => route.path === "/feedback");
